+++ packages/app/src/languages/actions/language-actions.ts
@@ -1,5 +1,5 @@
+import { ProgrammingLanguage } from "@ct/shared-types";
 import axios from "axios";
-import { ProgrammingLanguage } from "shared-types";
 import { ApiUrl } from "../../common/utils/backend-url";
 import { LanguagesStore } from "../stores/languages-store";