+++ packages/backend/src/megalinter/parsers/parse-sbom.ts
@@ -1,9 +1,9 @@
-import { SbomPackage } from "shared-types";
+import { SbomPackage } from "@ct/shared-types";
 import { fetchPackages } from "../../sbom/sbom-fetching-utils";
 import { Component, Dependency } from "../../sbom/sbom-types";
 import {
+  createSBOMPackages,
   extractMappingForApplication,
-  getPackages,
 } from "../../sbom/sbom-utils";
 import { ReportStore } from "../../stores/be-report-store";
 import { LinterCompleteMessage } from "../megalinter-types";
@@ -19,6 +19,7 @@ export const parseSBOM = async (
       msg?.outputSarif?.runs.length > 0 &&
       msg.outputSarif.runs[0]?.properties?.megalinter?.sbom
     ) {
+      reportStore.set({ fetchingSBOMPackages: true });
       console.log("parsing sbom...");
 
       const components = msg.outputSarif.runs[0].properties.megalinter.sbom
@@ -26,16 +27,17 @@ export const parseSBOM = async (
       const dependencies = msg.outputSarif.runs[0].properties.megalinter.sbom
         .dependencies as Dependency[];
       const applications = extractMappingForApplication(components);
-      const pkgsInfo = await fetchPackages(dependencies, components);
-      const sbomPackages: SbomPackage[] = await getPackages(
+      const rawPackages = await fetchPackages(dependencies, components);
+      const sbomPackages: SbomPackage[] = createSBOMPackages(
         dependencies,
         components,
         applications,
-        pkgsInfo
+        rawPackages
       );
       if (sbomPackages && sbomPackages.length > 0) {
         reportStore.set({ packages: sbomPackages });
       }
+      reportStore.set({ fetchingSBOMPackages: false });
     }
   }
 };